raise warning instead of IndexError for interpolation going beyond reach
what to do when interpolation goes beyond acceptable range?
old: raise an IndexError, because what you do makes no sense, but with the additional info how to make the interpolation anyway (by changing the default behavior...)
problem: the user will change the behavior to ignore this error and then won't reverse decision when finished with current analysis. That's not the point!
new: do not stop evaluation, clip value, but return a warning that what user does is a bad idea.
This way they can decide themselves what to do about the interpolation. Especially since there is a warning for every instance. I.e. the user gets more information than shutting down the program at the first encounter of an unsound value.